+++ app/src/main/java/com/infinity_coder/divcalendar/presentation/calendar/models/ChartPresentationModel.kt
@@ -1,10 +1,11 @@
 package com.infinity_coder.divcalendar.presentation.calendar.models
 
 import com.example.delegateadapter.delegate.diff.IComparableItem
+import java.math.BigDecimal
 
 data class ChartPresentationModel(
     val annualIncome: String,
-    val annualYield: Double,
+    val annualYield: BigDecimal,
     val currentCurrency: String,
     val monthlyPayments: List<MonthlyPayment>,
     val colors: List<Int>,
